Fixes #46 - Allow URLs to be built prior to the execution of the request.
Added a new `BuilderExtensions` file that builds things (duh). Within this file, add two new extension methods:
	* BuildUri - builds the URI and returns it as the Uri type.
	* BuildUriAsString - builds the URI but returns it as a string.

Added tests for the new methods.

Updated namespaces of some other tests.
