+++ src/Baseline.FluentHttpExtensions.Tests/Unit/BuilderExtensionsTests/BuildUriTests.cs
@@ -1,0 +1,19 @@
+using Xunit;
+
+namespace Baseline.FluentHttpExtensions.Tests.Unit.BuilderExtensionsTests
+{
+    public class BuildUriTests
+    {
+        [Fact]
+        public void It_Successfully_Builds_And_Returns_A_Uri()
+        {
+            var response = "https://www.google.com/"
+                .AsAGetRequest()
+                .WithPathSegment("search")
+                .WithQueryParameter("q", "how much does a pug snore")
+                .BuildUri();
+
+            Assert.Equal("https://www.google.com/search?q=how+much+does+a+pug+snore", response.ToString());
+        }
+    }
+}
