+++ src/Baseline.FluentHttpExtensions.Tests/Unit/SendTriggeringExtensionsTests/EnsureSuccessStatusCodeTests.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Moq;
 using Xunit;
 
-namespace Baseline.FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
+namespace Baseline.FluentHttpExtensions.Tests.Unit.SendTriggeringExtensionsTests
 {
     public class EnsureSuccessStatusCodeTests : UnitTest
     {
